requirements.txt
Django==1.10.6
filemagic==1.6
psycopg2==2.7
wheel==0.24.0
